UI: Fade In Active Editor Border Highlight
Our themes have two colors that we use as outlines for the editors, so
that we can differentiate the active area. This PR just fades in the
"Active Editor Outline" color rather than show it instantly. It makes
the change a little less jarring. As this transitions in the old active
area similarly transitions out. Note that this is very fast, only 150
milliseconds, so barely noticeable. Just a "softening" of the effect.

Pull Request: https://projects.blender.org/blender/blender/pulls/140836